requirements.txt
async-timeout==4.0.3
awscli==1.32.103
backports.tarfile==1.1.1
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
idna==3.7
jaraco.classes==3.4.0
jaraco.context==5.3.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
spotipy==2.23.0
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.12
urllib3==2.2.1
voluptuous==0.14.2
websocket-client==1.8.0
yolo==0.3.2
